show skills completed
